refactor : 객체 단일 책임에 따른 LottoProfitRate 분리
