+++ src/main/java/lotto/domain/LottoProfitRate.java
@@ -1,0 +1,33 @@
+package lotto.domain;
+
+import java.util.EnumMap;
+import java.util.Map.Entry;
+
+public class LottoProfitRate {
+
+    private static final int DECIMAL_TRIM_NUMERATOR = 100;
+    private static final double DECIMAL_TRIM_DENOMINATOR = 100.00;
+
+    private final double profitRate;
+
+    public LottoProfitRate(LottoResult lottoResult, int purchasedLottoPiece) {
+        this.profitRate = calculateProfitRate(lottoResult.getLottoResultStatistics(),
+            purchasedLottoPiece);
+    }
+
+    private double calculateProfitRate(EnumMap<LottoRank, Integer> lottosResultStatistics,
+        int lottoPiece) {
+        double sum = 0;
+        for (Entry<LottoRank, Integer> keyValue : lottosResultStatistics.entrySet()) {
+            sum += keyValue.getKey().getPrizeMoney() * keyValue.getValue();
+        }
+        double investCapital = lottoPiece * Lotto.LOTTO_PRICE;
+        double rawProfitRate = sum / investCapital;
+        return Math.round(rawProfitRate * DECIMAL_TRIM_NUMERATOR) / DECIMAL_TRIM_DENOMINATOR;
+    }
+
+    public double getProfitRate() {
+        return this.profitRate;
+    }
+
+}
